refactor: extract parser module
