+++ lua/haskell-tools/lsp/hover.lua
@@ -8,7 +8,7 @@
 
 local log = require('haskell-tools.log')
 local lsp_util = vim.lsp.util
-local HtUtil = require('haskell-tools.util')
+local HtParser = require('haskell-tools.parser')
 local OS = require('haskell-tools.os')
 local project_util = require('haskell-tools.project.util')
 
@@ -120,7 +120,7 @@ function hover.on_hover(_, result, ctx, config)
   _state.commands = {}
   local func_name = vim.fn.expand('<cword>')
   ---@cast func_name string
-  local _, signatures = HtUtil.try_get_signatures_from_markdown(func_name, result.contents.value)
+  local _, signatures = HtParser.try_get_signatures_from_markdown(func_name, result.contents.value)
   for _, signature in pairs(signatures) do
     table.insert(actions, 1, string.format('%d. Hoogle search: `%s`', #actions + 1, signature))
     table.insert(_state.commands, function()
