+++ tests/parser_spec.lua
@@ -1,4 +1,4 @@
-local util = require('haskell-tools.util')
+local HtParser = require('haskell-tools.parser')
 
 describe('Util:', function()
   it('Can get function signatures from LSP markdown docs', function()
@@ -12,7 +12,7 @@ describe('Util:', function()
       .. '```haskell\n'
       .. '_ :: forall a. Num a => a -> a\n'
       .. '```\n'
-    local func_sig, all_signatures = util.try_get_signatures_from_markdown('someFunc', doc)
+    local func_sig, all_signatures = HtParser.try_get_signatures_from_markdown('someFunc', doc)
     assert.same('Num a => a -> a', func_sig)
     assert.same({
       'Num a => a -> a',
